[CB] Feat : 파일 다운로드 기능 추가
  - 작성된 보도자료 파일을 볼 수 있도록 메서드 및 컨트롤러 추가

+++ src/main/java/com/gongyeon/io/netkim/model/jwt/JwtUtil.java
@@ -28,12 +28,12 @@ public class JwtUtil {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("USER_ROLE", String.class);
     }
 
-    public String createToken(int memberIdx, String userRole) {
+    public String createToken(long memberIdx, String userRole) {
         // 유효 시간 : 1시간 (1000 -> 1초 * 60 (1분) * 60 (1시간))
         Date exp = new Date(System.currentTimeMillis() + 1000*60*60*24*15);
         return Jwts.builder().
                 header().add("typ", "JWT")
-                .and().claim("memberIdx", Integer.toString(memberIdx))
+                .and().claim("memberIdx", Long.toString(memberIdx))
                 .claim("USER_ROLE", userRole)
                 .expiration(exp).signWith(secretKey)
                 .compact();
